requirements.txt
tensorflow
tensorflow-hub
numpy
opencv-python-headless
streamlit
Pillow
